Add healthz-cron URL for health checks of update tasks
This will be a check of the cron process running along side of the
web server process in the same container. Since there is one running
in every container using something like Dead Man's Snitch isn't very
useful since several could be down and the pings would still come in.